UPDATED / REMADE the request/response part of the server!
The way the server reads, parses, handles, and respondes to client
requests has been updated to make it cleaner and easier to scale up and
maintain. The main.cpp file is much cleaner and it easier to follow the
flow of the program.

Everything related to the requests/responses is handled inside the
src/server/ folder.

It's not easy to register new routes, and the code logic each of them
should follow can be placed inside src/server/routes/.

The server now also supports queries in the URI, and a request body for POST
requests.